Refactor broadcast channel selector
Decouple the inner broadcast channel selector from the automation
setting, so it can be reused out of the automation setting.
